% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{color_matrix}
\alias{color_matrix}
\title{Convert numeric matrix to color (character) matrix}
\usage{
color_matrix(x, colors = heat.colors(10))
}
\arguments{
\item{x}{Numeric or logical matrix.}

\item{colors}{Character vector of colors.}
}
\value{
Character matrix.
}
\description{
Given a numeric matrix, assign to each cell a color (character) value based on linearly interpolating a given vector of colors.
}

