% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{colour_matrix}
\alias{colour_matrix}
\title{Convert numeric matrix to colour (character) matrix}
\usage{
colour_matrix(x, colours = heat.colors(10))
}
\arguments{
\item{x}{Numeric or logical matrix.}

\item{colours}{Character vector of colours.}
}
\value{
Character matrix.
}
\description{
Given a numeric matrix, assign to each cell a colour (character) value based on linearly interpolating a given vector of colours.
}

