% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{inches_tall}
\alias{inches_tall}
\title{Get number of inches high a putative heatmap will be}
\usage{
inches_tall(x, row_lines = 1)
}
\arguments{
\item{x}{Character matrix of colours}

\item{row_lines}{Numeric value determining number of lines width each row of the heatmap should occupy.}
}
\value{
Numeric value.
}
\description{
Get number of inches high a putative heatmap will be
}

