% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{inches_wide}
\alias{inches_wide}
\title{Get number of inches across a putative heatmap will be}
\usage{
inches_wide(x, col_lines = 1)
}
\arguments{
\item{x}{Character matrix of colours}

\item{col_lines}{Numeric value determining number of lines width each column of the heatmap should occupy.}
}
\value{
Numeric value.
}
\description{
Get number of inches across a putative heatmap will be
}

