% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\docType{data}
\name{lines_between_hm_and_labels}
\alias{lines_between_hm_and_labels}
\title{Lines of space between the heatmap and row/column labels}
\format{An object of class \code{numeric} of length 1.}
\usage{
lines_between_hm_and_labels
}
\description{
Lines of space between the heatmap and row/column labels
}
\keyword{datasets}

