% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\docType{data}
\name{margin_lines}
\alias{margin_lines}
\title{Lines of space at margins of paintmap}
\format{An object of class \code{numeric} of length 1.}
\usage{
margin_lines
}
\description{
Lines of space at margins of paintmap
}
\keyword{datasets}

