% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{paintmap}
\alias{paintmap}
\title{Plot paintmap}
\usage{
paintmap(x, add = FALSE, ...)
}
\arguments{
\item{x}{Character matrix of colours}

\item{add}{Add ink to current viewport.}

\item{...}{Other graphical parameters for the rectangles of the grid to pass to \code{grid} function \code{gpar}, in turn passed to \code{grid} function \code{grid.rect}.}
}
\value{
Plots heatmap.
}
\description{
Plot paintmap
}
\examples{
paintmap(matrix(heat.colors(9), 3, 3, dimnames=list(letters[1:3], letters[4:6])))
}

